update RDS broker release to dev release including PG95 upgrade changes
Remove `postgis` as a default extension for Postgres 9.5 plans

We want to allow tenants to upgrade Postgres 9.5 database services to Postgres
10 or 11. It isn't possible to perform this upgrade succesfully when the
`postgis` extension is installed.

The RDS broker does not support disabling default extensions. Thus, to enable
the upgrade path, we remove `postgis` as a default extension, so that tenants
can disable it and upgrade. This will have the added benefit that newly created
Postgres 9.5 databases will automatically be able to upgrade.