Add concourse task to check cert hierarchy
Our certificate rotation process relies on having a simple hierarchy of
root CA signs leaf cert.

We do not have root CAs that sign intermediate CAs that sign leaf certs.

If we had an intermediate CA, our rotation process might break, because
we would rotate the root and the intermediate and the leaf at the same
time.

Add a task to the check-certificates jobs in the pipeline, which will
stop the pipeline if we have intermediate CAs

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>