New tool for tracking organisation quota upgrades
This commit adds a small new tool named `org_quota_history`. When
invoked using the instructions in its source code, it will output a
CSV of the orgs that exist, when/if they were upgraded to be billable,
etc.

Data is sourced from `paas-billing`'s `cfstore`. That stores a record
of the quotas PaaS orgs have had since we started recording data in
mid-2018.

Ideally this should be put in a better place than `paas-cf` but that's
left for later.