Ensure VPC peering CIDR blocks don't overlap with our infrastructure
We are going to allow GOV.UK to set up a VPC peering which is outside of the
range we normally grant to tenants. This will introduce a wider set of CIDR
blocks being alloated to tenants, in a way that isn't wholly consistent. To
prevent future us from accidentally allocating a block that overlaps with our
own infrastructure, this commit adds a test to check that very thing.