Use `bash` instead of `sh` in `register-csls-splunk-broker` task
We silenced Shellcheck rule SC2039 because it alerts on us using features that
aren't defined in the POSIX spec, but we still wanted to use those features.

But we were still using the `sh` shell to run the task, so the behaviour was
still undefined, and it still didn't work. Using `bash` ought to fix that.