Fix HTTP healthcheck in CF CLI v7 manifests
We configure many of our apps to use HTTP healthchecks. There is a
change in how this has to be configured which only became apparent
when deploying a new PaaS dev environment.

In CF CLI v6, we specify `health-check-type: http` in the app manifest
and CF will automatically health check the `/` route.

In CF CLI v7, the endpoint to be used for HTTP healthchecks now has to
be specified (there is no default.)

This only became apparent when Miki deployed a new environment because
in all existing environments CF has probably stored the default
endpoint value from when we still used CF CLI v6.