Add exception to the test
GOV.UK would like to have that specific subnet available to them, in
order to prevent any tearing down and spinning back up of the service.

We've discussed in the team, and find that plausible, but need to change
this test.

We're composing a list of ranges that we expect to be intact and only
checking these ranges in our test.