Bump RDS broker
We are not prescribing the patch level in the RDS service catalog plans in order to make use of the auto minor version upgrade feature.

This means that AWS chooses the patch level when creating or modifying an instance. If the patch-level chosen for the modification is for some reason lower than the patch-level of the current instance cf update-service will fail.

This change preserves the patch-level of the current instance when updating the plan, if the major and minor versions are the same. It also adds tests.

See https://github.com/alphagov/paas-rds-broker/pull/118 for further info