Add Terraform variable files to specific Terraform configurations
As of 0.12.x, defining a value for an undeclared variable causes a warning
during `terraform apply`.

> Using a variables file to set an undeclared variable is deprecated and will
> become an error in a future release. If you wish to provide certain "global"
> settings to all configurations in your organization, use TF_VAR_...
> environment variables to set these instead.

We receive this warning because we declare variables in
`terraform/cloudfoundry/variables.tf`, but define values for them in
`terraform/{env}.tfvars` When we run a configuration like `terraform/cyber`,
Terraform can see a value for a variable, but no declaration.

To work around this, we keep truly global variables in `globals.tf`, and
extract local variable values to (for example)
`terraform/cloudfoundry/{env}.tfvars`. Concourse is updated to supply these
`.tfvars` files.