Fix the stemcells not in sync alert
We have an alert that warns us if we are running multiple Stemcells
(OS base images.) This alert was hardcoded to expect 3 BOSH
Deployments.

We recently deployed the CF Autoscaler using BOSH. It is a separate
BOSH deployment, our fourth. When that was deployed the "multiple
stemcell" alert went off.

This commit edits that alert so that it directly counts the number of
stemcells in use, rather than being hacky. Credit to @tlwr.