Alert if the platform bills are not increasing
We recently broke `paas-billing-collector` for a few days, during
which the `paas_billing_total_costs_pounds` metric did not increase
at all. We were not very effectively alerted and it was pure luck that
we noticed before the end of the month.

In that situation I'd expect all the other billing alerts in this file
to go off, but I did not get any emails from them. While that's a bit
concerning, to begin with I just want to explicitly alert on bills not
increasing. That will warn us if this happens again.