Ignore CATS guids in tests
In 52b98b9 we've added a test to find any non-unique IDs to prevent some
issues in billing.

This test, is also validating weather the GUID is an actual GUID. This
sadly is not true, for any of our stubs and tests that create a GUID in
a form of: `CATS-2-SVC-PLAN-c73c869d65f9762e`.

To prevent the pipeline from erroring out, we can skip the GUID if it
starts with `CATS-`.