sqs-broker: register broker for specific orgs
The SQS broker will initially be restricted to a smaller subset of orgs
while in an alpha/beta period.

We register the broker in the admin org in all environments. This allows
us to ensure the registering logic is always working in dev/staging
environments

We register other orgs by adding their definition in the json files
found at `config/service-brokers/sqs/$DEPLOY_ENV-config.json`

For now; files exist to enable the broker for the `gds-tech-ops` org in
both `prod` and `prod-lon` environments.

Co-Authored-By: Philip Potter <philip.potter@digital.cabinet-office.gov.uk>
Co-Authored-By: Alex Monk <alex.monk@digital.cabinet-office.gov.uk>