sqs-broker: adds platform-tests for sqs lifecycle
extends the healthcheck app so that it responds to an sqs-tests endpoint
which uses binding credentials from VCAP_SERVICES to:

* connect to each queue (there is a "main" and a "dead-letter" queue for
  each provisioned service)
* try to send a message to the queue(s) with some attributes
* try to recv a message from the queue(s)

tests are added to the broker-acceptance suite to ensure that the
expected plans exists, that a queue can be provisioned, bound and used
(via the healthcheck app).

since this is the first of these brokers that requires async binding,
some additional polling logic has been added after the bind/unbind steps
to prevent a race.

The output from the cf-cli when polling for binding progress is
unfortunatly a bit ambiguous (things like "succeeded" can show up
multiple times as service-status and binding-status are conflated into
this one status command), so we use the presence of the bound
application name in the output as a marker for success which should be
"good enough".

Co-Authored-By: Philip Potter <philip.potter@digital.cabinet-office.gov.uk>
Co-Authored-By: Alex Monk <alex.monk@digital.cabinet-office.gov.uk>