Ship app logs to Logit from our monitoring apps
We recently had a 20-hour gap in the reporting of one of our custom
Elasticache metrics. It took several redeploys to fix that and we
don't know why. We also don't have any logs, because they've gone out
of the CF log buffer.

This commit starts shipping logs to Logit from the apps in the
`monitoring` space:

  - `paas-metrics`
  - `cloudwatch-exporter`
  - `paas-prometheus-endpoint-redis`

That will help us debug things in future.

Weird thing about this commit is that there's now a dependency between
some of the little deploy jobs, as only one of them creates the actual
syslog drain service. That seems to be the best thing to do for now.