Allow setup time before failing `paas-metrics`
This commit adjusts `paas-metrics`'s acceptance tests. Until now many
of them have expected the metric to be available immediately.

`paas_op_apps_count` was expected to be available immediately,
but I've seen it be unavailable for ~4 minutes when `paas-metrics`
starts. That's fine.

After this commit:

  - Every metric only has to eventually be there

  - Eventually only times out after 10 minutes

This should make it much less likely that we get annoying failures
where the metrics do eventually show up.