Configure SQS broker to only permit access from local IPs
Depends on https://github.com/alphagov/paas-sqs-broker/pull/15 and a
corresponding boshrelease to push that out.

This adds a new managed policy, restrict_to_local_ips, which can be
attached to an IAM User or Role to restrict use of that User/Role to
only be permitted from the egress IPs for the local VPC.  This
mitigates against credentials that are leaked.

The S3 broker has a similar feature in
s3_broker_ip_restriction_policy_arn, but that policy is scoped to only
apply to `s3:*` actions, rather than all actions.

I deliberately made the restrict_to_local_ips policy's name generic
because it could easily be used as-is for other brokers in future.