Link monitoring dashboard alerts count to Grafana not Prometheus
Our monitoring dashboard has a count of the currently firing alerts from Alert
Manager. It is configured to be a link to Prometheus, where one can see which
alerts are firing.

However, getting access to Prometheus is annoying, because it uses basic auth,
and to get those credentials we have to go to Credhub.

Instead, this commit changes the link to point to the same data, but presented
by Grafana. The user will already be logged in to Grafana, so it won't be
annoying.