Improve error message when not on the VPN
Until now, running commands such as `make dev pipelines` when not on
the VPN resulted in an awful and incorrect error:

  An error occurred (AccessDenied) when calling the ListAccountAliases
  operation: User: arn:aws:sts::_____:assumed-role/____________/______
  is not authorized to perform: iam:ListAccountAliases on resource: *
  with an explicit deny
  Required AWS account is dev, but your aws-cli is using keys for
  Makefile:16: recipe for target 'check-env' failed
  make: *** [check-env] Error 1

This commit detects when the AWS API call fails, and warns about that.
It also suggests users check they are on the VPN:

  An error occurred (AccessDenied) when calling the ListAccountAliases
  operation: User: arn:aws:sts::_____:assumed-role/____________/______
  is not authorized to perform: iam:ListAccountAliases on resource: *
  with an explicit deny

  Error talking to the AWS API. Check that you're using the VPN.
  make: *** [check-env] Error 1

This should stop users (including me!) getting hopelessly confused.
I also slightly improved the original error message, which happens
when using the wrong credentials (rare since we introduced aws-vault
and the GDS CLI.)