Further scale Gorouter to 15 instances
We may get an unpredictable spike in traffic soon. This commit scales us to 15 gorouters. This is a big increase from the previous 6, and intended to remove Gorouters as a bottleneck.

In Ireland we have run 12 Gorouters for some time without problems. London has more apps but I am hopeful we can run 15 without problems until we're clearer about how much traffic is coming our way.