Allow to pass a default into get_secrets
Add an optional argument to the get_secrets function in order to pass a
default (dummy) value to be used if the lookup in paas-credentials
fails.

Some secrets, such as the mailchimp API key, are only required in prod,
but we still need to set a dummy value in staging and dev environments as
it is expected to be non-nil during deployment/execution.