Clarify just how "free" service plans are
Service plans have a boolean property `free` which tells CloudFoundry if the
plan is available to trial orgs. Unfortunately, `cf marketplace -e SERVICE`
says "free" whenever that property is true. That's a bit misleading for
billable orgs, who DO have to pay for it.

We don't have influence over the format of the `cf marketplace` command, so
instead of correcting it be more nuanced, we introduce a clarifying statement
in the description of the "free" plans.