Run BBS on a higher-CPU instance type
For some time the active `diego-api` VM in Ireland has been exhausting its CPU credits. This VM runs BBS, a Cloud Foundry component which manages the cluster of app instances [1]. This has predictably experienced higher load as tenants have been doing more autoscaling.

This commit changes the EC2 instance type used for the `diego-api` VM. It upgrades from a t3.medium instance type to an m5.large` Benchmarks show that the m5.large's constant CPU performance is roughly equal to the burst performance of the t3.medium [2]. This means that it will have substantially more CPU available at all times.

Cost estimate is something like $120/month in each non-dev environment.

[1] https://docs.run.pivotal.io/concepts/diego/diego-architecture.html#components
[2] https://www.drlinkcheck.com/blog/t3-vs-m5



The motivation for this is that `diego-api` in Ireland has been exhausting its CPU credits for quite a long time. It appears to be caused by increased load on BBS