Ensure that all Bosh-generated certs have alternate names
In Go 1.15, treating the CommonName field of an x.509 certificate as
a host name when no Subject Alternative Name is present is
deprecated[1]. As a result, cf-deployment >= v14.0.0 has added SAN's
to all of their certificates.

To keep in step, we are now ensuring that all of the leaf
certificates we generate also have them.

Setting `update_mode: converge` on certificate type variables will cause
them to be regenerated immediately when there is any change to their
properties. In this instance, that will be a safe operation, because the
certificates will have the same authority, and the change is additive
only.

We might want to think about removing the option after this change has
rolled out, so that we don't accidentally shoot ourselves in the
foot in future.

[1] https://golang.org/doc/go1.15#commonname