Triple cloud controller rate limit in Ireland
There is a bug in the currently deployed version of Cloud Controller (1.102.0)
which is causing CF CLI v6 to misbehave and send polling requests as fast as it
possibly can. That's resulting in Notify getting rate limited when trying to
deploy.

We don't have a fix for the bug at the moment, and the rate limit is blocking
Notify's deploys. This commit triples the limit in Ireland to allow them to get
stuff out while we look for a fix.