isoseg: Respect the vm_tyme from data file
At the moment, this value seems to be ignored, in favour of some custom
functionality in regards to `isolation_segment_size`.

This is rather unwanted and perhaps unfamiliar behaviour, but we're not
here to fix the way we run the show, only the thing at hand.

Adding vm_type variable before the block, will make sure it has been set
to something, before the custom logic kicks in and perhaps overrides it.

The assignment line has been moved out of the block, to run after, to
make sure both cases are handled in the same way...

Ruby foo bad - sorry future generations.