Add notify production isolation segment
Mimics the segment given to their staging environment after load tests
on the staging segment have seen a significant impact in terms of
performance.

The plan is hopefully to only keep these around for a few months and pay
the cost of that whilst we bring things under control and can look at
reducing our CPU usage in a more long term manner.