Add 12 cells to Ireland
Notify are ditching their staging isolation segment for the time being. We need to restore the 
number of cells their staging environment was occupying before, so there's room for their 
apps to come back. We don't need to restore the full 24 from the isolation segment, because
they won't be running load tests agains staging and thus won't need the extra capacity.