Alter PSN Terraform to output multiple subnet group seeds
Cloud Foundry's Cloud Controller expects that security groups
will specify a destination as a single string. As a result, our
script for generating ops files for security groups bakes in
the assumption that its input will have a single string in the
corresponding part (the `subnet_cidr` parameter).

The PSN Terraform code was outputting an input for that script
which gave the `subnet_cidr` parameter as an array of strings,
because the PSN VPC connection has multiple IPs. This caused an
error when deploying to Cloud Foundry that we could only catch
in a production deploy.

To mitigate this, we mutate the output from the PSN Terraform
to output one seed per subnet. The script we have for
generating security groups is already capable of generating
multiple security groups. This is the safer approach, because
the VPC peering security groups generation script is more
widely used.