Silence CF app alerts by pointing them in to a black hole
The Prometheus Bosh release gives us a package of CloudFoundry
alerts. Many of these we care about, but two in particular are
very noisy and don't provide us any useful information:

  - CFAppCrashed
  - CFAppUnhealthy

The Bosh release doesn't feature configuration options for
which alerts to enable/disable, or to configure their alert
level.

Instead of making a PR to the upstream, which will take a long
time to get merged and released, we can reduce our alert noise
quickly by directing those alerts in to a black hole in
Alertmanager.