Fix setting max_allowed_packet for MySQL 5.7
A tenant using Drupal has had problems with the default value of
`max_allowed_packet` in RDS. The default is 4MB and that seems to
prevent using strings or blobs larger than that.

We tried to increase this to 256MB in the RDS Broker codebase [1].
That worked, but then broke again.

It turns out that MySQL 5.7 only is also configured in Terraform. It's
a bit tricky to try and remove it (needs some state editing.) For now
I'm just going to set `max_allowed_packet` in Terraform as well.

[1] https://github.com/alphagov/paas-rds-broker/pull/135