Prepend silenced alerts to Alertmanager routes config
The previous attempt at silencing some alerts had appended a configuration for
sending alerts in to an empty receiver to the list of rules. This did not work,
because the rules are processed in order, and an earlier rule direced the alert
to the warning reciever.

Instead, we now prepend the rules for the ones we want to silence, so that they
never reach the other receivers and are properly silenced.

This is still a workaround for being able to configure the alert level/which
alerts fire on the Prometheus bosh release.