Move AWS leaked key check into separate group
This is done as it prevented smoke tests etc being run in prod (only one
running job per serial group is allowed).
The leaked key detection job is just a simple bash script, so it is safe
to execute in parallel to smoke tests.

We also saw the checks taking more time in prod (17 mins) as compared to
dev (4 mins), so increasing the leaked key timer interval to 20 mins.