Prevent persistent dev environments being auto-deleted
Our development environments are usually torn down overnight as a cost
saving measure. We're now implementing two shared development environments
that we want to make persistent. To avoid them being configured to be torn
down, we introduce two new environment targets: `dev01` and `dev02`.

The Make syntax in this is perhaps a little arcane.

    .PHONY: dev01
    dev01: $(eval export DEPLOY_ENV=dev01)
    dev01: dev
        $(eval export ENABLE_AUTODELETE=false)
	@true

Declaring the `dev01` target twice will cause them to be run in the order
they're defined. In this case, the first one will run to set the
DEPLOY_ENV environment variable, and then the second one will run to
gather all the settings, and override ENABLE_AUTODELETE.