Revert "Fix persistant devs targets"
This reverts commit 689da944fb0096c6bb7194dbababe25cd8101244.

The commit is being reverted because defining it all in one Make task causes
the `dev` task dependency to be run before the `DEPLOY_ENV` environment
variable can be overriden.

The ordering important because the `dev` task defines a bunch of strings, like
the system URL, which in dev envs contain the value of the `DEPLOY_ENV`
environemnt variable. If it isn't overriden first, those strings will contain
whatever the value of `DEPLOY_ENV` is on the developer's machine, instead of
`dev01` and `dev02`.