Don't upgrade auto-scaler's stemcells to bionic
We have found that the bionic upgrade is damaging the autoscaler in a
way, where the metricserver is being unresponsive, meaning the scaling
cannot be happening due to 500 errors from unreachable BOSH DNS
internal connection.

Reverting it back in the hope of it bringing back the autoscaler -
feature used by our tenants and hoping the upstream will perform more
tests.