Buildpack email generation interactively prompts for highlights
When we send out an email to tenants about buildpack upgrades, we list all of
the dependency version changes, along with links to all of the release notes.
This is decent chunk of information, and it would be easy for tenants to not
read through all of the release notes for all the buildpacks they use, and/or
to miss something important.

To make life easier for them, we should read the release notes ourselves as
operators, and summarise the important bits (e.g important versions being
removed, shiny new versions being added, whole new buildpacksb being
introduced).

The buildpack email generation Go program now interactively prompts users to
read the notes and edit a message about them (n.b. it respects the $EDITOR env
var). Those messages are compiled in to the final email.