Upload stemcells during autoscaler deploy
As in [1] and [2], the CF app autoscaler isn't working with the Ubuntu Bionic
stemcells right now, and needs to be deployed with the old Xenial stemcells
instead.

This has been working in environments that existed prior to the Xenial->Bionic
switch, because the Xenial stemcell image continued to exist in Bosh's storage.
However, this isn't true for new environments, and thus they fail to deploy the
autoscaler.

To remedy this, we take a copy of the stemcell uploading Concourse task, and
pop it in the pipeline just before the autoscaler is deployed. We also
replicate the ops file that sets the stemcell from `cf-manifest`. The two
combine to download the correct stemcell image, and upload it to Bosh.

[1] https://github.com/cloudfoundry/app-autoscaler-release/pull/254
[2] https://github.com/cloudfoundry/app-autoscaler-release/issues/253