Exclude infra subnets from AZ disabling
The infra subnet hosts all of our Bosh and Concourse related
VMs and databases. Usually, this is zone A, but the Concourse
database lives in a random AZ within the region. Disabling the
AZ in which Concourse's database lives (when that's different
from the VMs) forces Concourse to die unexpectedly.

To work around that, we exclude it from being black holed. It
isn't important for CloudFoundry to operate.

As an added bonus, disabling the zone in which Bosh and/or
Concourse live should no longer stop them working.