create-cloudfoundry pipeline: disable bosh resurrector before disabling VPCs
when we disable an AZ at the network level it causes the resurrector
to spring into action. the problem with this being that it can never
succeed and while attempting to do so it will hold onto locks that
are needed to e.g. send another deployment through.

we should probably consider adding an alert to remind an operator to
re-enable the resurrector once an incident is over.