create-cloudfoundry pipeline: add resurrector-is-enabled job
the idea of this job is to both give a visual indication of the current
resurrector status and generate periodic alerts when it is disabled,
gently reminding operators to re-enable it.

the alerts should be easily disable-able through
ENABLE_ALERT_NOTIFICATIONS whilst keeping the visual indication
functional, which should make it useful for e.g. dev environments