create-cloudfoundry pipeline: add az-*-is-enabled-in-manifest jobs
as with resurrector-is-enabled, the idea of these jobs is to both give a
visual indication of the current manifest status and generate periodic
alerts when an AZ is disabled, gently reminding operators to re-enable
it.

the alerts should be easily disable-able through
ENABLE_ALERT_NOTIFICATIONS whilst keeping the visual indication
functional, which should make it useful for e.g. dev environments