create-cloudfoundry pipeline: add az-*-is-enabled-in-vpc jobs
as with resurrector-is-enabled, the idea of these jobs is to both give a
visual indication of the current vpc status and generate periodic
alerts when an AZ is disabled, gently reminding operators to re-enable
it.

deliberately not pulling the actual id of the network ACL from the
terraform statefile in an attempt to be able to detect ACLs that have
been orphaned or clickops'd. however we still rely on the ACL being
named in a particular way, so it's not perfect. but i'm not going to
write a full ACL parser/emulator that would be needed for more
accurate detection.

the alerts should be easily disable-able through
ENABLE_ALERT_NOTIFICATIONS whilst keeping the visual indication
functional, which should make it useful for e.g. dev environments