Align log-api instance count with dopplers
> Traffic Controller resources are usually scaled in line with Doppler
> resources. The recommended formula for determining the number of Traffic
> Controller instances is:

> Number of Traffic Controller instances = Number of Doppler instances / 2

As we do not work in single instances, we increase things in pairs of threes.

We're also refining the tests for number of dopplers... As we seem to be
just on the edge and to incorporate the following:

> Doppler resources can require scaling to accommodate your overall log
> and metric volume. The recommended formula for estimating the number of
> Doppler instances you need to achieve a loss rate of < 1% is:

> Number of Doppler instances = doppler.ingress / 16 000
> where doppler.ingress represents the per-second rate of change of the
> ingress metric from the doppler source, summed over all VMs in your
> deployment.

Now, this isn't as simple as our rule of thumb ratio 2:1 with cells. But
is fairly close...

After examining the logs, we're just on the edge after recent changes,
missing that one extra doppler...

Along with the refined tests, I'm increasing the counts.