Raise London API rate limit to 60,000
In commit 6f7a23e601f4 we raised the API rate limit in Ireland to 60k because
of a bug. In London, we're now experiencing a different bug that could be
alleviated by raising the API rate limit.

Specifically, we're seeing tenants getting rate limited when using
`paas-prometheus-exporter` to monitor large orgs.