healthcheck: Add simple containerized endpoint
We'd like to be simply running a small application on different set of
services, with very minimal response time and data. We'd like to assure
the response is not cachable - or rather that we can determine if it
hasn't been.

It consists of a timestamp, so the client can determine when the
response has been made _really_.