Add links for new instance group jobs
As a result of changes introduced in the latest version of app-autoscaler some new links are introduced, these target cf as the deployment. We don't use the default name of cf in order to avoid collisions in the same AWS account, as a result we need to override it with an ops file.