NLB listener for PaaS secrets mTLS must be listen for TCP only
When using a TLS listener, TLS is terminated there and potentially
re-established on the next hop to the target group. When that happens, the
client certificate is not presented to the server, and the server's certificate
is not presented to the client (instead the LB listener's certificate is).

To get around that, we use a TCP listener so that the packets are forwarded
between the two sides, and TLS is not interfered with.