Update for testing billing in both regions
As billing costs differently in different regions we need to be able to
test both.

This means it makes sense to put the tests here. The acceptance tests
will pull in the correct config and features for the region it is run
in.

It order to get the tests running easily locally, I modifed the rates to
no longer have epochs as that was messing up when testing locally due to
beginning of month constraints.