Fix billing formulae and tests
There was a problem with the billing for rds instances in that they
charged for a whole months of usage even if only a single hour was used.

This was due to the ceil function being called on a variable.

The fix is to remove the ceil function and to update the tests to verify
this.

The spreadsheets that are used to calculate the tests have also been updated.