Correct path to TF state file in AZ healthcheck output extraction
The path to the state file was wrong, which caused an error message to be
silently written to the output file, which was subsequently ingested in a later
task. The value it ingested was of no use to it, and caused an error, which
manifest itself as a failed healthcheck.