Update dev* make targets and add branch target
* Rather than hardcoding "devxx" as targets, use a wildcard rule.
* Add `current-branch` target, that populates the `BRANCH`
  environment variable as the current checked out branch

Implementation note:

Previously, `SYSTEM_DNS_ZONE_NAME` and `APPS_DNS_ZONE_NAME` were set
twice, once in `dev:` and once in the individual environment target (ie.
`dev01:`).
By escaping the `$` for the DEPLOY_ENV variable in `dev:`, it is not
expanded at the time the make target runs, but instead by the shell
at runtime. This allows a little bit of deduplication!

[skip ci]