Update makefile help generation
Rather than the old way of just pulling out the targets, categorise and
sort the targets into categories and then alphabetically sort the
targets within these groups. This allows targets to be identified and
found easier.

This script isn't intended to be run manually - it is the default target
in the Makefile, so is only run if running `make` with no arguments, or
`make help`