Refactor Github Actions workflow to re-use a workflow
We have a handful of groups of tests that we want to run against the
repository, and they all share the setup steps for the environment. In the
past, Github Actions didn't support any kind of task re-use, so we made do with
a matrix to run the groups of tests as different configurations.

GHA now supports task re-use, so we should refactor the workflow to properly
represent each set of tests as a different job with a shared workflow.