Scale infra for cell increase in iso-segments
The use of isolation segments bypasses some of our checks to ensure
scaling of infra components proportional to diego-cell counts.

We are adding a total of 36 cells (12+24) for Notify who intend to do
load testing and expect a general rise in requests. This will raise the
total cell count in Ireland to 94.

In order to ensure the platform is not bottlenecking Notify, we scale
infra to roughly map to ratios in prod-lon.