Run make smoke for billing smoke tests
The new billing work introduces database tests into the acceptance
tests.

As it stands this doesn't make much sense for smoke testing. 'make smoke' in
the nbi-part-two branch (the branch that should be merged to make the
new billing code live), change 'make smoke' to exclude the database
tests and be appropriate for running api tests.