Add an OAuth client for the API for use by PaaS admin local development
We want to be able to point our local PaaS admin instances at a CloudFoundry
development environment, so that we don't have to push to CF all the time when
working on it.

WARNING: The OAuth client is configured insecurely. It has
1. static credentials
2. a callback URL of "localhost:3000".
3. full admin access.

This is not something that should be present outside of development
environments. A warning has been added to the generate manifest script.