Combine loops, handle deletion failure
Pulled the processing loop into the org fetching loop as in prod the
argument list got too long when collated the list of all orgs.
Also added in handling of orgs failing deletion so that it does not
break execution. Names of orgs failing deletion are written to a file.