Upgrades the TLS cipher suite we use on load balancers
Upgrades from to 'ELBSecurityPolicy-TLS-1-2-2017-01' to
'ELBSecurityPolicy-TLS-1-2-Ext-2018-06' for ALB/NLB and
'ELBSecurityPolicy-TLS-1-2-2017-01' for classic load balancers. As it stands
today (2022-02-23), AWS does not have an ELB security policy which supports TLS
1.3 on Application Load Balancers. One is avaibale for Network Load Balancers,
but in the interests of consistency we will only use it when ALBs have the same
support.

The policy names for ALB/NLBs and classic load balancers have diverged in AWS,
so we've had to diverge the two properties here too.

The presence of the 'SslPolicy' field on the 'ModifyListener' action of AWS'
Elastic Load Balancing API [1] suggests that we will be able to modify the
policy without downtime. Equally, the AWS Terraform provider documentation
makes no mention of needing to destroy the listener before making the change.

[1] https://docs.aws.amazon.com/elasticloadbalancing/latest/APIReference/API_ModifyListener.html