Mitigate 'app app' pricing plan being created
We are having a problem with our calculator in ireland

We narrowed this down to a pricing plan called 'app app' that is being
created, by generateMissingPlans in paas-billing. For some reason it
thinks there is not the correct pricing plan and is generating one.

This turns off all pricing plan generation. Which may cause issues as
the collector needs to have pricing plan for each event it has. An
alternate mitigation will be suggested.