add an initial simplified log-cache instance group
this lays the ground for cf-deployment 19.0.0 to go out.

cf-deployment >=18 separates the log-cache from the doppler instances.
it helps if all the cluster's instances are aware of the log-cache
bosh dns alias before the full deployment goes out.

we're going for the same number of instances and vm types as the
existing dopplers so we know we're not underprovisioning. once
we've observed production load using this we can decide what to
downscale and how.

doing the scaling & sizing in this initial deployment mostly so
the tests don't whine at us about something it sees as a
misconfiguration.