manifest generation: add a reordering mechanism
it is useful to be able to enforce a specific ordering for some
arrays on the final manifest we pass to bosh. this allows maps
to specify __paas_order keys which will be used in a final
sorting of the entries. entries/arrays without this key should
be unaffected.

the specific use case for this initially is to denote a specific
update order of the instance_groups.