cdn-broker: provision a proper uaa secret, include in secret rotation
for some reason we have been providing a secret for the cdn broker
which has insufficient permissions to authenticate in the way the
broker was trying to. switch to a configuration a lot more like
everything else.

also move the provisioning of this secret to 330-uaa.yml along with
all the other uaa secrets. this is probably why it got missed in the
rotation routine. add it to rotate-cloudfoundry-credentials.

change name of uaa secret from cdn_broker to cdn-broker to convince
everything to switch over smoothly and also match the other secrets'
naming.