Enable overlapping blocks in Prometheus' timeseries database
We are going to be introducing some new recording rules to improve query
performance, for which we will need to backfill the metric data. In order to do
that, Prometheus' timeseries database needs to be allowed to have overlapping
blocks of time, so that it can slot in the new metric data during compaction.

This commit ought to be reverted once the new recording rules have been added,
and the metric data backfilled.