Skip alternative_name check for an upstream cert
The context for why is listed here: https://github.com/cloudfoundry/cf-deployment/pull/985

I have made a specific test that the alternative_names == 0 so that this
change will be remembered and reverted when upstream is fixed.

Once that PR is merged and the alternative name is added, the if
statement can be removed and the tests will stop failing.