Enable buildpack upgrades to be targeted
There are some instances where we'd like to upgrade just one or two buildpacks,
but the tool previously would update all of the ones found in the configuration
file.

The tool now takes a list of buildpacks that should be updated, and will update
those in the list and maintain the current configuration for the others.

E.g.
  ./scripts/update_buildpacks.sh python_buildpack nginx_buildpack